timestamp/requirements.txt
streamlit
youtube-transcript-api
rake-nltk
nltk
requests
google-generativeai
beautifulsoup4
lxml
pytube
